Feature/shops

+++ app/Http/Livewire/Package/ShowPackages.php
@@ -4,6 +4,7 @@ namespace App\Http\Livewire\Package;
 
 use App\Models\Package;
 use App\Traits\StorePackage;
+use Illuminate\Support\Facades\Auth;
 use Livewire\Component;
 use Livewire\WithFileUploads;
 use Illuminate\Validation\Rules\File;
@@ -16,6 +17,8 @@ class ShowPackages extends Component
 
     public $searchEmail = '';
 
+    public $searchShopname = '';
+
     public $searchLastname = '';
 
     public $searchStatus = '';
@@ -26,9 +29,9 @@ class ShowPackages extends Component
 
     public $searchCity = '';
 
-    public $sortField = 'created_at';
+    public $sortField = 'shops.name';
 
-    public $sortDirection = 'desc';
+    public $sortDirection = 'asc';
 
     public function sortBy($field)
     {
@@ -43,12 +46,16 @@ class ShowPackages extends Component
     public function render()
     {
         return view('livewire.package.show-packages', [
-            'packages' => Package::search('lastname', $this->searchLastname)->
-            search('email', $this->searchEmail)->
-            search('status', $this->searchStatus)->
-            search('streetname', $this->searchStreetname)->
-            search('zipcode', $this->searchZipcode)->
-            search('city', $this->searchCity)->
+            'packages' => Package::select('packages.*', 'shops.name')->
+            join('shops', 'shops.id', '=', 'packages.shop_id')->
+            whereIn('shops.id', Auth::user()->shops->pluck('id'))->
+            search('packages.lastname', $this->searchLastname)->
+            search('packages.email', $this->searchEmail)->
+            search('packages.status', $this->searchStatus)->
+            search('packages.streetname', $this->searchStreetname)->
+            search('packages.zipcode', $this->searchZipcode)->
+            search('packages.city', $this->searchCity)->
+            search('shops.name', $this->searchShopname)->
             orderBy($this->sortField, $this->sortDirection)->paginate(10),
         ]);
     }
@@ -87,6 +94,7 @@ class ShowPackages extends Component
             $package->housenumber = $packageArrayToAdd[4];
             $package->zipcode = $packageArrayToAdd[5];
             $package->city = $packageArrayToAdd[6];
+            $package->shop_id = $packageArrayToAdd[7];
 
             $errors = $this->StorePackage($package);
         }
@@ -94,6 +102,8 @@ class ShowPackages extends Component
 
     public function delete(Package $package)
     {
-        $package->delete();
+        if (in_array($package->shop_id, Auth::user()->shops->pluck('id')->toArray())) {
+            $package->delete();
+        }
     }
 }
